fix(web): lazy-split the composer routes so tiptap leaves the entry chunk (#2523)
`@kampus/composer` (tiptap/ProseMirror) shipped in the SPA's single entry chunk
to every public route (pano, sözlük, landing, /mecmua index), gating first paint
on a heavy editor almost no one loads — a performance-pillar violation (ADR 0162).

The issue's premise that /lab/composer is the sole composer consumer is stale:
#2544/#2549 added the mecmua editor as a second consumer, and MecmuaIndexPage
eagerly imported the gate helpers off MecmuaEditorPage, dragging tiptap into the
entry chunk regardless of any route split. Root-cause fix:

- Extract the composer-free gate core (`mecmuaPublishAffordance`,
  `shouldShowMecmuaWriteCta`, `MecmuaPublishAffordance`) into a new module
  `apps/web/src/pages/mecmua-write-gate.ts`; repoint MecmuaIndexPage and the
  unit test there. Gate semantics unchanged — only the helpers relocate.
- Lazy-split both composer route components in App.tsx behind React.lazy +
  Suspense (Turkish "yükleniyor…" fallback): /lab/composer → LabComposerPage,
  and /mecmua/yaz + /mecmua/yaz/:id → MecmuaEditorPage.

Proof (pnpm build): tiptap moves from the 1,329.84 kB entry chunk (before) to a
separate async chunk (after: entry index-*.js 827.77 kB with zero tiptap matches;
the composer/prosemirror payload in src-*.js loaded only via dynamic import from
the route chunks, never preloaded on first paint).

Fixes #2523
